requirements.txt
#python>=3.6
numpy>=1.16.4
scipy>=1.3.0 
pandas>=0.24.2 
mayavi>=4.6.2
scikit-learn>=0.21.2
opencv>=3.4.2
PyYAML>=5.1.1
